+++ src/main/java/com/geoffrey/betting/sofascorelocal/domain/provider/ConnectorMode.java
@@ -1,5 +1,0 @@
-package com.geoffrey.betting.sofascorelocal.domain.provider;
-
-public enum ConnectorMode {
-    LOCKED_OFFLINE_J1
-}